Update README.md
If tilde.club is going to be the central repository for unix servers being run for social purposes, tildecabal seems redundant. The scripts repo here should not be for general interest. Everything here is specifically for totallynuclear.club. All general server documentation should be collected at tilde.club's github account, which seems to be their intention. 

I removed the reference to "tildeboxes" because I think "Unix servers for social purposes" is an adequate descriptor. I don't want to perpetuate the idea that this is some kind of branded network (and the wording "unix server for social purposes" is taken directly from the tilde.club github org page).